fix(comment-automation): send ai agent private replies on instagram channels
processCommentAIReply only dispatched private DMs when channelType was
"messenger" (stale "no private_replies API" comment, predating the
Instagram private-reply support added in #874). instagram and
instagramFacebook silently no-op'd: no send, no throw, no log line
(issue #863). Wires up the already-working
sendInstagramLoginPrivateReply/sendInstagramFacebookPrivateReply
instead, mirroring executePrivateReply's existing three-channel
dispatch.

Also splits the ~830-line comment-automation/index.ts into focused
files (automation-matching, hide-comments, public-reply, private-reply)
and extracts a shared CommentAutomationChannelType, since the
private-reply dispatch table would otherwise need to be duplicated
between index.ts and ai-reply.ts. Adds test coverage for the
instagram/instagramFacebook AI-reply private-DM paths and corrects a
comment-attachment.ts docstring that wrongly implied parity with
executePrivateReply's channel support.

Refs #863
